+++ PlotApp/MVVM/Models/Function/Function.cs
@@ -3,7 +3,7 @@ using PlotApp.MVVM.Models.Dot;
 
 namespace PlotApp.MVVM.Models.Function {
     internal class Function {
-        public Function(IEnumerable<Point> points, double scaleX, double scaleY, double wrapX, double wrapY, string name = "") {
+        public Function(IEnumerable<Point> points, decimal scaleX, decimal scaleY, decimal wrapX, decimal wrapY, string name = "") {
             this.Points = new List<Point>(points);
             this.ScaleX = scaleX;
             this.ScaleY = scaleY;
@@ -13,10 +13,10 @@ namespace PlotApp.MVVM.Models.Function {
         }
 
         public string      Name   { get; set; } = string.Empty;
-        public double      ScaleX { get; set; } = 1.0;
-        public double      ScaleY { get; set; } = 1.0;
-        public double      WrapX  { get; set; } = 0.0;
-        public double      WrapY  { get; set; } = 0.0;
+        public decimal      ScaleX { get; set; } = 1.0M;
+        public decimal      ScaleY { get; set; } = 1.0M;
+        public decimal      WrapX  { get; set; } = 0.0M;
+        public decimal      WrapY  { get; set; } = 0.0M;
         public List<Point> Points { get; set; } = new ();
     }
 }
